various, mainly adding SIZEOF(...) and BITS(...) expressions

+++ quxlang/include/quxlang/data/class_layout.hpp
@@ -11,8 +11,8 @@ namespace quxlang
     struct class_layout
     {
         std::vector<class_field_info> fields;
-        std::size_t size = 0;
-        std::size_t align = 0;
+        std::uint64_t size = 0;
+        std::uint64_t align = 0;
 
         RPNX_MEMBER_METADATA(class_layout, fields, size, align);
     };
